feat: write tensor utils
