+++ jev4_cosine_results.md
@@ -1,0 +1,5 @@
+# Cosine Similarity Matrix
+
+| Item | I:beach.jpg |
+|---|---|
+| I:beach.jpg | 1.000 |
